requirements.txt
certifi==2018.10.15
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
Django==2.1.2
django-cleanup==2.1.0
django-cors-headers==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.8.2
djangorestframework-simplejwt==3.2.3
drfapikey==0.0.3
idna==2.7
itypes==1.1.0
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.0
openapi-codec==1.3.2
Pillow==5.3.0
Pygments==2.2.0
PyJWT==1.6.4
pytz==2018.5
PyYAML==3.13
requests==2.19.1
simplejson==3.16.0
uritemplate==3.0.0
urllib3==1.23
virtualenv==16.0.0
